chat-agent/requirements.txt
openai>=1.3.0,<1.4.0
python-dotenv==1.0.0
snowflake-connector-python==3.15.0
pandas==2.1.4
langgraph==0.0.34
pydantic==2.5.0
typing-extensions>=4.12.0
sqlalchemy 
snowflake-sqlalchemy
fastapi==0.109.2
uvicorn==0.27.1
numpy>=1.20.0
requests==2.31.0